refactor: Separate aapt injection
